Separate constants from main helper
Separate constants and other classes from the main WinSyscallsHelper.cs
